+++ src/Microsoft.Azure.WebJobs.Host/Bindings/BindingProviders/FilteringBindingProvider.cs
@@ -41,8 +41,7 @@ namespace Microsoft.Azure.WebJobs.Host.Bindings
 
             var attr = context.Parameter.GetCustomAttribute<TAttribute>();
 
-            var clonerCtx = AttributeClonerContext.New(_nameResolver, context.Parameter);
-            var cloner = new AttributeCloner<TAttribute>(attr, context.BindingDataContract, clonerCtx);
+            var cloner = new AttributeCloner<TAttribute>(attr, context.BindingDataContract, _nameResolver);
             var attrNameResolved = cloner.GetNameResolvedAttribute();
 
             // This may do validation and throw too.
