+++ src/Microsoft.Azure.WebJobs.Host/WebJobs.Host.csproj
@@ -467,7 +467,6 @@
     <Compile Include="Bindings\BindingProviders\GenericCompositeBindingProvider.cs" />
     <Compile Include="Bindings\BindingProviders\ItemBindingProvider.cs" />
     <Compile Include="Bindings\BindingProviders\ValidatingWrapperBindingProvider.cs" />
-    <Compile Include="Bindings\AttributeClonerContext.cs" />
     <Compile Include="Bindings\SysBindingData.cs" />
     <Compile Include="Bindings\DefaultAttributeInvokerDescriptor.cs" />
     <Compile Include="Bindings\FuncArgumentBuilder.cs" />
